Fixed port for spotDL docker-compose
This should solve issue #60. As noted by issue https://github.com/spotDL/spotify-downloader/issues/2035 of the original project, at the moment there seems sto be an issue with the port used by the container, as it (wrongly) connects to port 8800 regardless of the port specified when looking for songs. A user suggested (and I personally verified that it works on my installation) changing the port to 8800, and it does indeed fix the issue.